backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
google-genai>=1.0.0
